+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Usernames [#574](https://github.com/rokwire/core-building-block/issues/574)
+- Use signature Key ID to check specific key for service account auth [#481](https://github.com/rokwire/core-building-block/issues/481)
+
+### Fixed
+- Service registration error handling change [#468](https://github.com/rokwire/core-building-block/issues/468)
 
 ## [1.24.2] - 2022-08-08
 ### Added